Facu Pauro: agrego modo oscuro en configuracion y ajustes
